+++ src/i18n/i18n.ts
@@ -5,7 +5,11 @@ export var MAP = {
     'MAP_BASE_LAYERS': 'Base layers',
     'MAP_LAYERS': 'Layers',
     'MAP_WORKSPACE_LAYERS': 'Workspace',
-    'SEARCH_PLACEHOLDER': 'Enter a place to go...'
+    'SEARCH_PLACEHOLDER': 'Enter a place to go...',
+    'CONTEXTMENU.CENTER_MAP':'Center map here',
+    'CONTEXTMENU.ZOOM_IN':'Zoom in',
+    'CONTEXTMENU.ZOOM_OUT':'Zoom out',
+
 
   },
 
@@ -14,7 +18,11 @@ export var MAP = {
     'MAP_BASE_LAYERS': 'Capas base',
     'MAP_LAYERS': 'Capas',
     'MAP_WORKSPACE_LAYERS': 'Espacio de trabajo',
-    'SEARCH_PLACEHOLDER': 'Introduce un lugar...'
+    'SEARCH_PLACEHOLDER': 'Introduce un lugar...',
+    'CONTEXTMENU.CENTER_MAP':'Centrar el mapa aquí',
+    'CONTEXTMENU.ZOOM_IN':'Acercar',
+    'CONTEXTMENU.ZOOM_OUT':'Alejar',
+
   }
 
 };